ClienteAdapter done

+++ src/ClienteComum.java
@@ -8,7 +8,7 @@ public class ClienteComum extends Cliente  {
 
     }
 
-    public ClienteComum(String nome, String email, String celular, String senha, float saldo, Carrinho carrinho, String endereco, boolean assinaturaAtiva, String validadeAssinatura) {
+    public ClienteComum(String nome, String email, String celular, String senha, double saldo, Carrinho carrinho, String endereco, boolean assinaturaAtiva, String validadeAssinatura) {
         super(nome, email, celular, senha, saldo, carrinho, endereco, assinaturaAtiva, validadeAssinatura);
     }
